reverse

+++ 0x05-pointers_arrays_strings/4-print_rev.c
@@ -6,7 +6,7 @@
 
 void print_rev(char *s)
 {
-	int long i;
+	int long long i;
 
 	while (s[i] != '\0')
 	{
